get rid of crazy code duplication with _buffer_proxy
it still should be scrapped, since we have other routines doin similar
things but it's whatever right now
